Remove Float from decomposers

+++ test/decomposer/test_cnot_decomposer.py
@@ -7,7 +7,7 @@ import pytest
 from opensquirrel.decomposer.cnot_decomposer import CNOTDecomposer
 from opensquirrel.decomposer.general_decomposer import check_gate_replacement
 from opensquirrel.default_gates import CNOT, CZ, SWAP, H, Ry, Rz, X
-from opensquirrel.ir import ControlledGate, Float, Gate
+from opensquirrel.ir import ControlledGate, Gate
 
 
 @pytest.fixture(name="decomposer")
@@ -15,10 +15,7 @@ def decomposer_fixture() -> CNOTDecomposer:
     return CNOTDecomposer()
 
 
-@pytest.mark.parametrize(
-    ("gate", "expected_result"),
-    [(H(0), [H(0)]), (Rz(0, Float(2.345)), [Rz(0, Float(2.345))])],
-)
+@pytest.mark.parametrize(("gate", "expected_result"), [(H(0), [H(0)]), (Rz(0, 2.345), [Rz(0, 2.345)])])
 def test_ignores_1q_gates(decomposer: CNOTDecomposer, gate: Gate, expected_result: list[Gate]) -> None:
     check_gate_replacement(gate, expected_result)
     assert decomposer.decompose(gate) == expected_result
@@ -49,10 +46,4 @@ def test_CZ(decomposer: CNOTDecomposer) -> None:
     gate = CZ(0, 1)
     decomposed_gate = decomposer.decompose(gate)
     check_gate_replacement(gate, decomposed_gate)
-    assert decomposed_gate == [
-        Rz(1, Float(math.pi)),
-        Ry(1, Float(math.pi / 2)),
-        CNOT(0, 1),
-        Ry(1, Float(-math.pi / 2)),
-        Rz(1, Float(math.pi)),
-    ]
+    assert decomposed_gate == [Rz(1, math.pi), Ry(1, math.pi / 2), CNOT(0, 1), Ry(1, -math.pi / 2), Rz(1, math.pi)]
